refactor(local-fs): isolate Windows drive discovery
Keep Windows-specific detection behind the existing filename-based CI classification overlay and avoid changing protected gate configuration.
