+++ scripts/ci/change-impact.json
@@ -122,7 +122,6 @@
         "src/main/crash-diagnostics.ts",
         "src/main/index.ts",
         "src/main/lifecycle-shutdown.ts",
-        "src/main/local-fs/service.ts",
         "src/main/local-rpc-transport.ts",
         "src/main/notebook/environment-discovery.ts",
         "src/main/notebook/environment-state-tracker.ts",
